+++ js/plugins/clickToHighlight.js
@@ -1,6 +1,10 @@
 /**
- * DataFormJS [clickToHighlight] Plugin
+ * DataFormJS [clickToHighlight] Plugin.
  * 
+ * This allows a user to easily see where they are on
+ * wide table rows or mobile devices or highlight other
+ * elements under a list.
+ *
  * This file depends on CSS, example:
  *     .highlight {background-color:yellow; }
  */
@@ -17,14 +21,21 @@
 (function () {
     'use strict';
 
-    function toggleHighlight() {
+    function toggleHighlight(e) {
+        // Do not change highlight when a link is clicked. The reason is because
+        // it causes a quick flash of the highlighted style on the row before
+        // the page changes. Since the user is clicking to another page rather
+        // than highlighting a row the style change is not desired.
+        if (e.target.nodeName === 'A') {
+            return;
+        }
         /* jshint validthis:true */
         this.classList.toggle('highlight');
     }
 
     var clickToHighlight = {
-        onRendered: function () {
-            var elements = document.querySelectorAll('.click-to-highlight');
+        onRendered: function (rootElement) {
+            var elements = (rootElement || document).querySelectorAll('.click-to-highlight');
             Array.prototype.forEach.call(elements, function(el) {
                 var n, m;
                 if (el.tagName === 'TABLE' && el.tBodies.length === 1) {
@@ -32,7 +43,7 @@
                     for (n = 0, m = rows.length; n < m; n++) {
                         rows[n].style.cursor = 'pointer';
                         rows[n].addEventListener('click', toggleHighlight);
-                    }                    
+                    }
                 } else {
                     for (n = 0, m = el.children.length; n < m; n++) {
                         el.children[n].style.cursor = 'pointer';